+++ ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: gkehren <gkehren@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/13 15:24:26 by gkehren           #+#    #+#             */
-/*   Updated: 2022/03/13 15:26:58 by gkehren          ###   ########.fr       */
+/*   Updated: 2022/03/19 16:13:03 by gkehren          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,12 +17,12 @@ char	*ft_stpncpy(char *dst, const char *src, size_t len)
 	int	i;
 
 	i = 0;
-	while (src[i] != '\0' && i < len)
+	while (src[i] != '\0' && i < (int)len)
 	{
 		dst[i] = src[i];
 		i++;
 	}
-	while (i < len)
+	while (i < (int)len)
 	{
 		dst[i] = '\0';
 		i++;
